fixed bugs for libs

+++ _lib/cpp/_src/macros.hpp
@@ -26,7 +26,7 @@ using vvpil = vector<vector<pil>>;
 #define tostr to_string
 
 #ifndef CONSTANTS
-    constexpr long long INF = 1e18;
+    constexpr ll INF = 1e18;
     constexpr int MOD = 1000000007;
     constexpr ld EPS = 1e-10;
     constexpr ld PI = M_PI;
